% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcor.R
\name{rpcor}
\alias{rpcor}
\title{rPCor for variable screening}
\usage{
rpcor(
  X,
  Y,
  block = FALSE,
  group.index.Y = NULL,
  alpha = 1e-05,
  xcor.thres = 0.8,
  ycor.thres = 0.8,
  max.reach = 5,
  d = 20,
  ht = FALSE,
  c = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{X:}{The n*p predictor matrix.}

\item{Y:}{The n*q response matrix.}

\item{block:}{Logical; is there any block/module structure in the response
variables? If \code{TRUE}, rPCor-module will be run. If \code{FALSE},
rPCor-correlation will be run. Default is \code{FALSE} and rPCor-correlation
will be run.}

\item{group.index.Y:}{Group information for response variables (Y). Only need to
be specified if \code{block=TRUE}.}

\item{alpha:}{The alpha threshold. Default is \code{1e-5}.}

\item{xcor.thres:}{The threshold for between-predictor (robust) correlation.
Default is 0.3.}

\item{ycor.thres:}{The threshold for between-response (robust) correlation. Only
need to be specified if \code{block=FALSE}. Default is 0.3.}

\item{max.reach:}{The maximum order for partial corrlation. Default is 5.}

\item{d:}{The maximum number of neighbors to be considered as candidates for
conditioning set. That is, for predictors \eqn{X_i}, only top d predictors that
have the largest marginal (robust) correlation with \eqn{X_i} will be considered
for \eqn{X_j}'s conditioning sets. Same for responses if \code{block=FALSE}.}

\item{ht:}{Logical; is the data heavy-tailed? If \code{TRUE}, robust partial
correlation will be calculated. Otherwise, regular partial correlation will be
used. Default is \code{FALSE}. Note that \code{ht=TRUE} will significantly
increase the computation burden.}

\item{c:}{Truncation parameter, which will further be used to find the
robustification parameter \code{gamma}. See the paper for details. The larger
the c, the more truncation. Default is 0.5.}

\item{verbose:}{Logical; should the results for each partial corrlation order
be printed out? Default is \code{TRUE}.}
}
\value{
A p*q 0-1 indicator matrix. Each row is a predictor and each column is
a response. For each element of the matrix, 1 means this predictor-response
pair is kept after screening while 0 means it's removed.
}
\description{
This function will run the \code{rPCor} to perform variable screening
on high-dimensional predcitor and high-dimensional response.
}
\examples{
\dontrun{
set.seed(1234)


}
}
